fixing tests for use long data type

+++ lib/src/main/java/com/sweetzpot/stravazpot/authenticaton/model/AppCredentials.java
@@ -2,19 +2,19 @@ package com.sweetzpot.stravazpot.authenticaton.model;
 
 public class AppCredentials {
 
-    private final int clientID;
+    private final long clientID;
     private final String clientSecret;
 
-    public static AppCredentials with(int clientID, String clientSecret) {
+    public static AppCredentials with(long clientID, String clientSecret) {
         return new AppCredentials(clientID, clientSecret);
     }
 
-    public AppCredentials(int clientID, String clientSecret) {
+    public AppCredentials(long clientID, String clientSecret) {
         this.clientID = clientID;
         this.clientSecret = clientSecret;
     }
 
-    public int getClientID() {
+    public long getClientID() {
         return clientID;
     }
